breaking out stimulus controller
