+++ app/components/feedback/feedback_modal_component.html.erb
@@ -7,7 +7,7 @@
     <% end %>
     <% component.with_body do %>
         <%= helpers.turbo_frame_tag "feedback-form_frame" do %>
-            <% form_with url: feedback_form_path, method: :post, id: "feedback-form", class: "feedback-form", role: "form", namespace: 'feedback', data: { 'controller': 'feedback-form', 'action': 'turbo:submit-end->feedback-form#showToast' } do |f| %>
+            <% form_with url: feedback_form_path, method: :post, id: "feedback-form", class: "feedback-form", role: "form", namespace: 'feedback', data: { 'controller': 'feedback-form feedback-modal', 'action': 'turbo:submit-end->feedback-modal#showToast' } do |f| %>
                 <div class="modal-body">
                     <%= render Feedback::FeedbackFormFieldsComponent.new(form: f) %>
                 </div>
