+++ app/components/feedback/feedback_standalone_component.html.erb
@@ -1,9 +1,9 @@
-<div class="standalone mx-auto">
+<div class="standalone mx-auto" data-controller="feedback-standalone" data-feedback-standalone-target="container">
     <h2 class="fw-normal"><%= t('blacklight.feedback_form.title') %></h2>
 
     <div class="feedback-form-container">
         <%= helpers.turbo_frame_tag "feedback-form_frame" do %>
-        <% form_with url: feedback_form_path, method: :post, id: "feedback-form", class: "feedback-form", role: "form", namespace: 'feedback', data: { 'controller': 'feedback-form', 'action': 'turbo:submit-end->feedback-form#hideForm' } do |f| %>
+        <% form_with url: feedback_form_path, method: :post, id: "feedback-form", class: "feedback-form", role: "form", namespace: 'feedback', data: { 'controller': 'feedback-form', 'action': 'turbo:submit-end->feedback-standalone#hideForm' } do |f| %>
                 <%= f.hidden_field :form_format, value: 'standalone' %>
                 <%= render Feedback::FeedbackFormFieldsComponent.new(form: f) %>
                 <div class="d-flex-inline text-end">
