+++ app/javascript/controllers/index.js
@@ -61,6 +61,12 @@ application.register("facet-list", FacetListController)
 import FeedbackFormController from "./feedback_form_controller"
 application.register("feedback-form", FeedbackFormController)
 
+import FeedbackModalController from "./feedback_modal_controller"
+application.register("feedback-modal", FeedbackModalController)
+
+import FeedbackStandaloneController from "./feedback_standalone_controller"
+application.register("feedback-standalone", FeedbackStandaloneController)
+
 import GalleryPreviewController from "./gallery_preview_controller"
 application.register("gallery-preview", GalleryPreviewController)
